# Conflicts:
#	test/CMakeLists.txt

+++ api/boostpython/api_routing.cpp
@@ -11,7 +11,7 @@ namespace expose {
     void routing_path_info() {
 
         py::class_<routing_info>("RoutingInfo","Describe the hydrological distance and the id of the target routing element (river)")
-         .def(py::init<py::optional<int,double>>(py::args("id","distance"),"create an object with the supplied parameters"))
+         .def(py::init<py::optional<int64_t,double>>(py::args("id","distance"),"create an object with the supplied parameters"))
          .def_readwrite("id",&routing_info::id,"id of the target,down-stream river")
          .def_readwrite("distance",&routing_info::distance,"the hydrological distance, in unit of [m]")
          ;
@@ -55,7 +55,7 @@ namespace expose {
             " This definition is recursive, and we use RiverNetwork to ensure the routing graph\n"
             " is directed and with no cycles.\n"
             )
-            .def(py::init<int,py::optional<routing_info,routing::uhg_parameter>>(py::args("id","downstream","parameter"),
+            .def(py::init<int64_t,py::optional<routing_info,routing::uhg_parameter>>(py::args("id","downstream","parameter"),
                 "a new object with specified parameters, notice that a valid river-id|routing-id must be >0"
                 )
             )
@@ -74,7 +74,7 @@ namespace expose {
 
     }
     void routing_river_network() {
-        routing::river& (routing::river_network::*griver)(int)= &routing::river_network::river_by_id;
+        routing::river& (routing::river_network::*griver)(int64_t)= &routing::river_network::river_by_id;
         py::class_<routing::river_network>("RiverNetwork",
             "A RiverNetwork takes care of the routing\n"
             "see also description of River\n"
